✨ 코어 데이터 완성
