+++ Alarmi/Alarmi/Sources/CoreData/CoreDataDate+CoreDataClass.swift
@@ -10,7 +10,6 @@
 import Foundation
 import CoreData
 
-
 public class CoreDataDate: NSManagedObject {
 
 }
