--- Alarmi/Alarmi/Sources/Models/CoreDataDate.swift
@@ -1,30 +1,0 @@
-//
-//  CoreDataDate.swift
-//  Alarmi
-//
-//  Created by 임 용관 on 2022/07/29.
-//  Copyright © 2022 MoTe. All rights reserved.
-//
-
-import Foundation
-import CoreData
-import UIKit
-
-//@objc(CoreDataGoalModel)
-//class CoreDataGoalModel: NSManagedObject {
-//    @NSManaged var id: UUID?
-//    @NSManaged var callDate: Date?
-//    @NSManaged var callTimePeriod: NSNumber!
-//    @NSManaged var start: Date?
-//    @NSManaged var end: Date?
-//}
-
-//class CoreDataMoteDate: NSManagedObject {
-//    @NSManaged var moTeDate: Date?
-//}
-
-//extension CoreDataDate {
-//    @nonobjc public class func fetchRequest() -> NSFetchRequest<CoreDataDate> {
-//        return NSFetchRequest<CoreDataDate>(entityName: "CoreDataDate")
-//    }
-//}
