+++ Alarmi/Alarmi/Sources/Register/RegisterPlanViewController.swift
@@ -96,10 +96,8 @@ final class RegisterPlanViewController: UIViewController {
 extension RegisterPlanViewController {
     @objc private func buttonDidTap() {
         
-//        planCoreDataSave(callTimePeriod: callTimePeriod, callDate: goal.startDate)
         updateCoreData(day: goal.startDate, keyName: "callDate")
         updateCoreData(period: Int16(callTimePeriod), keyName: "callTimePeriod")
-        updateMoteDate(day: goal.startDate, keyName: "moteDate")
         
         switch type {
         case .register:
